+++ test/Integration/Gateways/GpApiConnector/CreditCardNotPresentTest.php
@@ -77,6 +77,7 @@ class CreditCardNotPresentTest extends TestCase
         $this->assertNotNull($response);
         $this->assertEquals('SUCCESS', $response->responseCode);
         $this->assertEquals(TransactionStatus::CAPTURED, $response->responseMessage);
+        $this->assertNull($response->payerDetails);
     }
 
     public function testCreditSaleWithFingerPrint()