+++ test/Integration/Gateways/GpApiConnector/GpApiApmTest.php
@@ -331,7 +331,7 @@ class GpApiApmTest extends TestCase
         $order->hasInsurance = true;
         $order->description = 'Order description';
 
-        $response = $this->paymentMethod->charge(14)
+        $response = $this->paymentMethod->charge(29)
             ->withCurrency($this->currency)
             ->withDescription('New APM Uplift')
             ->withAddress($this->shippingAddress, AddressType::SHIPPING)